Arrumando os arquivos
- Criando o arquivo para as Structs
- Removendo o arquivo que estava usando para teste
- Add documentação interna do código
